optimizing logger outputs
