+++ src/utils/testconfig.ini
@@ -18,7 +18,7 @@ handlers=consoleHandler,fileHandler
 class=FileHandler
 level=DEBUG
 formatter=fileFormatter
-args=(__import__("datetime").datetime.now().strftime(f"src/utils/{__name__}_%%Y-%%m-%%d_%%H-%%M-%%S.log"),)
+args=(__import__("datetime").datetime.now().strftime(f"logs/{__name__}_%%Y-%%m-%%d_%%H-%%M-%%S.log"),)
 
 [handler_consoleHandler]
 class=StreamHandler
